Add files via upload
This is in no way a COMPLETE RSS parser. There were a lot of things I wanted to add, like an entire database, multiple sources, etc, but on account of my limited knowledge couldn't do so. So, if you guys can help me figure out on how I can integrate a database with my code, it'd be really appreciated!